+++ sdk/assemblyscript/scripts/upgrade-deps.sh
@@ -11,6 +11,16 @@ cd src
 ncu -u -t minor
 npm install
 
+cd ../templates
+for template in *; do
+	if [[ -d ${template} ]]; then
+		cd "${template}"
+		ncu -u -t minor
+		npm install
+		cd ..
+	fi
+done
+
 cd ../examples
 for example in *; do
 	if [[ -d ${example} ]]; then